requirements.txt
fastapi
uvicorn
mangum
Pillow
numpy
pydantic
aiohttp
uvloop
gunicorn
aiofiles
turicreate
boto3
beautifulsoup4
requests
bs4
google-search-results
lxml
joblib
python-dotenv
